Data Storage and Browser Swapping
Implemented a button on the browser bar to swap search engines as a dropdown menu, using QWidgetActions to space text and images effectively. The browser can now be swapped at will, though I need to redesign the swap system to not redirect you to the search engine homescreen once selected.

Also added the framework for an SQL based data storage system. Currently in plain text files but will be updated to encrypt information