Updates to Customisation Features
Added improvements to the button layout alongside additional invert features.

Added tabs

Added a right click menu option that allows for the adjustment of buttons on a colour system with either custom or presets.

Colours currently cannot be saved but the infrastructure is in place for implementation

The right click menu appears for bot buttons and on the main screen. Buttons currently do not work but there is functional CSS injection that allows for webpage background colour changes. Next steps for this feature is to allow for profile saving for automatic injection and specific selections for different parts rather than just right click to adjust specifically the background

Refactored components of the codebase to prepare for later additions and include improvements to the primary systems for development in button and ui customisation

notes and updates for this project will likely be less in-depth since it's only the basic rapid development method for browser testing. A proper browser will be developed in a different repository with a superior upload schedule.