Rebuild
Rebuilding the browser customisation system and marking project test 2 as a failure. UI element selection with right click is too unnecessarily complicated and results in massively overblown issues due to recreated right click menus for different elements and constant issues in button tracking.

New customisation design will instead use an 'overlay' selector that greys out the screen to disable button inputs and allows users to select and apply colours to any button at any point. I may reattempt the previous strategy with the full JS/C++ version of this, but if it comes with the same problems as the python deployment, I hope that this becomes a more convenient backup option if implemented properly.